Make filter less aggressive w.r.t. pull response
Also split off into separate file as we are not backporting this.
