+++ gossip/src/lib.rs
@@ -7,6 +7,7 @@ pub mod contact_info;
 pub mod crds;
 pub mod crds_data;
 pub mod crds_entry;
+mod crds_filter;
 pub mod crds_gossip;
 pub mod crds_gossip_error;
 pub mod crds_gossip_pull;
